+++ google-books-application/src/app/core/auth/constants/navigation-page-data.const.ts
@@ -1,4 +1,0 @@
-export const NAVIGATION_PAGE_DATA = [
-  { name: 'Overview', url: '/books-overview' },
-  { name: 'Favorite', url: '/books-favorite' },
-];