+++ google-books-application/src/app/core/header/constants/navigation-page-data.const.ts
@@ -1,0 +1,6 @@
+import { APP_ROUTES } from '../../../shared/constants/app-routes.const';
+
+export const NAVIGATION_PAGE_DATA = [
+  { name: 'Search', url: APP_ROUTES.SEARCH },
+  { name: 'Favorite', url: APP_ROUTES.FAVORITE },
+];