+++ google-books-application/src/app/modules/books/books-routing.module.ts
@@ -1,18 +1,18 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
-import { BooksOverviewComponent } from './components/books-overview/components/overview-page/books-overview.component';
 import { BookDetailComponent } from './components/book-detail/components/detail-page/book-detail.component';
 import { AuthGuard } from '../../core/auth/guards/auth.guard';
 import { BooksFavoriteComponent } from './components/books-favorite/books-favorite.component';
+import { BooksSearchComponent } from './components/books-search/components/overview-page/books-search.component';
 
 const routes: Routes = [
   {
-    path: 'books-overview',
-    component: BooksOverviewComponent,
+    path: 'search',
+    component: BooksSearchComponent,
     canActivate: [AuthGuard],
   },
   {
-    path: 'books-favorite',
+    path: 'favorite',
     component: BooksFavoriteComponent,
     canActivate: [AuthGuard],
   },