+++ google-books-application/src/app/modules/books/constants/search-categories.const.ts
@@ -1,9 +1,0 @@
-export const SEARCH_CATEGORIES: string[] = [
-  'All',
-  'Art',
-  'Biography',
-  'Computers',
-  'History',
-  'Medical',
-  'Poetry',
-];