+++ google-books-application/src/app/modules/books/models/search-params.interface.ts
@@ -1,6 +1,4 @@
 export interface SearchParams {
   searchTerm: string;
-  category: string;
-  orderBy: string;
   startIndex: number;
 }