+++ google-books-application/src/app/shared/constants/app-routes.const.ts
@@ -1,6 +1,6 @@
 export const APP_ROUTES = {
-  LOGIN: 'login',
-  LOGOUT: 'logout',
-  BOOKS_OVERVIEW: 'books-overview',
-  BOOKS_FAVORITE: 'books-favorite',
+  LOGIN: '/login',
+  LOGOUT: '/logout',
+  SEARCH: '/search',
+  FAVORITE: '/favorite',
 };